AI_meeting_by_Gradio/requirements.txt
gradio>=5.0.0
pyaudio>=0.2.13
openai>=1.0.0
numpy>=1.20.0
python-dotenv>=1.0.0
requests>=2.28.0
pillow>=9.0.0
soundfile>=0.12.1
librosa>=0.9.2
pydub>=0.25.1
wave>=0.0.2
ollama>=0.1.5
